add intterfaces DAO
