+++ services/src/main/java/by/academy/it/rentacar/actions/PriceService.java
@@ -1,10 +1,10 @@
 package by.academy.it.rentacar.actions;
 
 import by.academy.it.rentacar.dao.PriceDAO;
-import by.academy.it.rentacar.entity.Fuel;
-import by.academy.it.rentacar.entity.Price;
 import by.academy.it.rentacar.enums.Transmission;
 import by.academy.it.rentacar.exceptions.DAOException;
+import by.academy.it.rentacar.entity.Fuel;
+import by.academy.it.rentacar.entity.Price;
 
 /**
  * Class PriceService induces PriceDAO
